ManifoldNormal.jl

This packages defines distributions and noises on manifolds. They are defined as follows:

A distribution is a fully prescribed probability distribution on a given manifold; an example is the ActionDistribution distribution.

A noise is a distribution on a manifold, but parameterized by a point on that manifold; examples are IsotropicNoise, ActionNoise and NoNoise.

Noise Models

A general interface for noises on manifolds.

Example Usage

Let us set up a simple manifold and a point on it.

using ManifoldsBase
using ManifoldNormal
σ = 2.0 # the standard deviation
M = ManifoldsBase.DefaultManifold(2)
# choose an isotropic noise for this example
noise = IsotropicNoise(M, σ)
x = rand(M)
""
# output
""

The sample space is the space where noisy points are expected.

sample_space(noise)
# output
DefaultManifold(2; field = ℝ)

The covariance of the noise at the point $x$. It is a covariance operator on the tangent space at $x$.

Σ = get_covariance_at(noise, x, DefaultOrthonormalBasis()) 
# output
2×2 PDMats.ScalMat{Float64}:
 4.0  0.0
 0.0  4.0

Using noise(rng, x) to produce sample from the noise distribution centred at x.

import Random
rng = Random.default_rng()
y = noise(rng, x) # another point of M
using Manifolds
is_point(M, y)
# output
true

Interface

AbstractNoise interface

All the noises are subtypes of the AbstractNoise abstract type. The following methods are expected:

sample_space

get_covariance_at

ManifoldNormal.get_basis_at

ManifoldNormal.add_noise

ManifoldNormal.rescale_noise

ManifoldNormal.add_noise — Function

add_noise(::AbstractNoise, ::Random.AbstractRNG, x::TX) :: TX

Add noise at point x.

Shortcut: noise(rng, x) ≡ add_noise(noise, rng, x)

ManifoldNormal.add_noise — Method

add_noise(::NoNoise, ::Random.AbstractRNG, x)

The identity function. No noise is added.

ManifoldNormal.get_basis_at — Function

get_basis_at(::AbstractNoise, x) :: AbstractBasis

The basis given by the noise model at the point x.

ManifoldNormal.get_covariance_at — Function

get_covariance_at(::AbstractNoise, x, ::AbstractBasis)  :: AbstractPDMat

A covariance matrix for this noise, at the given point, expressed in the given basis.

ManifoldNormal.get_covariance_at — Method

get_covariance_at(n::NoNoise, ::Any, ::Any)

Returns a zero covariance matrix, as this is a deterministic distribution.

ManifoldNormal.rescale_noise — Function

rescale_noise(TN<:AbstractNoise, λ) :: TN

Rescale the noise by a given factor $λ$. If $X$ is the random variable distributed as the noise, then return the distribution of $λX$.

Shortcut: λ * noise ≡ rescale_noise(noise, λ)

ManifoldNormal.sample_space — Function

sample_space(::AbstractNoise) :: AbstractManifold

Manifold on which the noise operates.

Methods

Action Noise

Consider a group action $G ⊂ \mathrm{Diff}(M)$. The push-forward of a normal distribution on a Lie algebra. The push-forward at a point $x\in M$ is done by the function $F : \mathfrak{g} \to M$ defined by

\[F(ξ) = \exp(ξ)⋅ x\]

ManifoldNormal.AbstractActionNoise — Type

A model of a map from a manifold to a covariance on a Lie algebra.

ManifoldNormal.ActionNoise — Type

ActionNoise(A::GroupAction, σ::Number)

Convenience method to create an action noise with a constant, isotropic covariance, by default with respect to the standard metric of the Lie algebra.

ManifoldNormal.ActionNoise — Type

ActionNoise(A::GroupAction, Σ::PDMat)

Convenience method to create an action noise with a constant covariance Σ, by default with respect to the standard scalar product of the Lie algebra.

ManifoldNormal.ActionNoise — Type

ActionNoise(
    action, # group action G ⊂ Diff(M)
    covariance, # function M -> AbstractPDMat
    basis # fixed basis of Alg(G)
)

Covariance defined on Lie algebra. The corresponding distribution centred at x₀ is the push forward of the normal distribution on the Lie algebra by the function ξ ↦ \exp(ξ)⋅x₀.

ManifoldNormal.adjoint_noise — Method

adjoint_noise(noise::ActionNoise{<:GroupOperationAction})

If the noise is defined by a group operation action, compute the corresponding noise for the action with switched sides. It is based on the identities

\[χ \exp(ξ) = \exp(χ ξ χ^{-1}) χ\]

and

\[\exp(ξ) χ = χ \exp(χ^{-1} ξ χ)\]

ManifoldNormal.get_adapted_basis_at — Method

Obsolete: in this basis, the covariance matrix is the identity.

ManifoldNormal.get_lie_covariance_at — Function

get_lie_covariance_at(
  ::AbstractActionNoise,
  x, # point in manifold M
  B, # AbstractBasis of Alg(G)
  ) :: AbstractPDMat

Covariance on Alg(G), where G is the action group.

ManifoldNormal.update_cov — Method

update_cov(n::ActionNoise, Σ)

New action noise with same action and basis but with new (possibly point dependent) covariance Σ.

Isotropic Noise

This noise at a point $x$ on a metric manifold $M$ is the push-forward by the metric exponential of a centred normal distribution on the tangent space $T_x M$. The covariance of that normal distribution is equal to the metric at that point $x$.

ManifoldNormal.IsotropicNoise — Type

IsotropicNoise(manifold, # metric manifold M
  deviation # function M -> 𝑹
)

Model a noise with covariance equal to the metric at the given point on the manifold.

ManifoldNormal.IsotropicNoise — Method

IsotropicNoise(M, std::Number)

Create an isotropic noise on the manifold M. The number std is then the standard deviation, which does not depend on the point of the manifold.

ManifoldNormal.isotropic_perturbation — Method

isotropic_perturbation(rng, M, p)

Create noisy vector at the point p on the metric manifold M, where the covariance matrix is the identity in an orthonormal basis.

No Noise

A convenient noise which is just the point-mass distribution at a point $x$ on any manifold.

ManifoldNormal.NoNoise — Type

NoNoise(M::AbstractManifold)

Models the absence of noise on the given sample space M.

Action Distributions

This is essentially an ActionNoise but with a fixed reference point.

ManifoldNormal.ActionDistribution — Type

ActionDistribution(μ, noise::ActionNoise)

Distribution stemming from an action noise specialized at one point.

ManifoldNormal.ActionDistribution — Type

ActionDistribution(action::Action, μ, Σ::PDMat, B::Basis)

Create an action distribution from the given action, mean and covariance.

ManifoldNormal.action_noise — Method

action_noise(D::ActionDistribution) :: ActionNoise

Create an action noise object from a ActionDistribution distribution. This is simply an action noise with constant covariance.

ManifoldNormal.get_action — Function

get_action(d::AbstractActionDistribution) :: AbstractGroupAction

The underlying action (i.e., the underlying homogeneous space).

ManifoldNormal.get_lie_basis — Function

get_lie_basis(d::AbstractActionDistribution) :: Basis

The basis of $\mathfrak{g}$ in which the covariance is defined.

ManifoldNormal.random_vector_lie — Method

random_vector_lie(
    rng::Random.AbstractRNG,
    d::AbstractActionDistribution) = begin

Draw a random vector in the Lie algebra following the covariance in the Lie algebra.

ManifoldNormal.scaled_distance — Method

scaled_distance(D::ActionDistribution, x)

Start with a ActionDistribution(μ,Σ) distribution with mean $μ$ lying on the sample manifold $M$, and covariance $Σ$ in the Lie algebra $\mathfrak{g}$ of the group $G$. The infinitesimal action of the group $G$ on the manifold $M$ gives rise to the linear map $P \colon \mathfrak{g} \to T_{μ}M$, which in turns gives the projected covariance $PΣP^*$ on the tangent space $T_{μ}M$.

One then measures the scaled distance from $μ$ to $x$ from $v := \log(μ, x)$ with the formula

\[\frac{\sqrt{v^T (PΣP^*)^{-1} v}}{\sqrt{n}}\]

(where $v$ is regarded as a column matrix here) and $n$ is the dimension of the sample manifold $M$.

ManifoldNormal.update_mean — Method

update_mean_cov(d::ActionDistribution, μ)

Return new ActionDistribution object with new mean $μ$.

ManifoldNormal.update_mean_cov — Method

update_mean_cov(d::ActionDistribution, μ, Σ)

Return new ActionDistribution object with new mean $μ$ and covariance $Σ$.